Retry when we recieve a 406 error.
The API server now uses HTTP status 406 to indicate that you've
specified an operation which depends on a resource which has not
yet made it to the created state (for example creating an
instance on a network which isn't built yet). This change adds
a retry for when the client sees that happen.